connect button when no wifi and changes to getting fruit information on click

+++ AndroidStudioProjects/Fruit/app/src/main/java/com/example/jojo/fruit/Fruit.java
@@ -39,7 +39,7 @@ public class Fruit {
         mWeight = weight;
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return mWeight;
     }
 }
